Fix some errors from tests failing

+++ src/taxpasta/infrastructure/application/motus/motus_profile_standardisation_service.py
@@ -42,7 +42,7 @@ class MotusProfileStandardisationService(ProfileStandardisationService):
 
         """
         result = profile.loc[~(profile[MotusProfile.tax_id].isnull() & (profile[MotusProfile.read_count]==0)),[MotusProfile.tax_id, MotusProfile.read_count]].copy()
-        result[MotusProfile.tax_id]= result[MotusProfile.tax_id].astype(int)
+        result[MotusProfile.tax_id]= result[MotusProfile.tax_id].astype(float).astype(int)
         result[MotusProfile.read_count]= result[MotusProfile.read_count].astype(int)
         result.columns = [StandardProfile.taxonomy_id, StandardProfile.count]
         return result
